Replace the class used for generating API specification fields marked as "uri"
As the GitHub API uses URITemplates {/replaceme} in the URIs, the default java.net.URI could not parse them.
The builtin Java URI class is now replaced with Spring's UriTemplate which parses the URIs with curly braces correctly, as well as maintains backwards compatibility with the Java standard URI class.