Add Retrofit testing
Adding Wiremock and Spock frameworks to allow writing component tests for the service.
Adding a Spring Boot test profile to reconfigure the GitHub API calls to the local Wiremock instance.
Introducing two test cases which uses two complete response payloads to mimic a success and not success path of the GitHub Search API.